created a common module and change the code accordingly
